Add files via upload
Question 1: Given an array of student's heights, how many students will have to swap in order to be in sorted order?
Question 2: Assume that a certain letter on your keyboard sticks, creating unknown number (including zero) extra instances of a letter with every press.  Given a person's name, and a string, is it possible that the string was an attempt to type that person's name?